I84: Ability to tag icons - BE
- Introduce the repository layer in the persistency (aka db) layer
