+++ backend/integration-tests/db/repositories/icon-repo/delete-icon-from-db.spec.ts
@@ -2,10 +2,11 @@ import { randomBytes } from "crypto";
 import { Pool } from "pg";
 import { flatMap, map} from "rxjs/operators";
 
-import { manageTestResourcesBeforeAndAfter, getCheckIconfile } from "./db-test-utils";
-import { Iconfile, IconfileDescriptor } from "../../src/icon";
-import { createIcon, deleteIcon, query, deleteIconfile } from "../../src/db/db";
-import { boilerplateSubscribe } from "../testUtils";
+import { manageTestResourcesBeforeAndAfter } from "../../db-test-utils";
+import { Iconfile } from "../../../../src/icon";
+import { query } from "../../../../src/db/db";
+import { boilerplateSubscribe } from "../../../testUtils";
+import { createIcon, deleteIcon } from "../../../../src/db/repositories/icon-repo";
 
 describe("deleteIconFromDB", () => {
 
@@ -13,19 +14,16 @@ describe("deleteIconFromDB", () => {
 
     it("should delete all entries associated with the icon", done => {
         const user = "zazie";
-        const iconfileDesc: IconfileDescriptor = {
-            format: "french",
-            size: "great"
-        };
         const iconfileInfo: Iconfile = {
             name: "metro-icon",
-            ...iconfileDesc,
+            format: "french",
+            size: "great",
             content: randomBytes(4096)
         };
 
         createIcon(getPool())(iconfileInfo, user)
         .pipe(
-            flatMap(() => deleteIconfile(getPool())(iconfileInfo.name, iconfileDesc, user)),
+            flatMap(() => deleteIcon(getPool())(iconfileInfo.name, user)),
             flatMap(() => query(getPool(), "select count(*) as row_count from icon", [])),
             map(result => expect(result.rows[0].row_count).toEqual("0")),
             flatMap(() => query(getPool(), "select count(*) as row_count from icon_file", [])),
