+++ backend/integration-tests/db/repositories/icon-repo/delete-iconfile-from-db.spec.ts
@@ -2,10 +2,11 @@ import { randomBytes } from "crypto";
 import { Pool } from "pg";
 import { flatMap, map} from "rxjs/operators";
 
-import { manageTestResourcesBeforeAndAfter } from "./db-test-utils";
-import { Iconfile } from "../../src/icon";
-import { createIcon, deleteIcon, query } from "../../src/db/db";
-import { boilerplateSubscribe } from "../testUtils";
+import { manageTestResourcesBeforeAndAfter } from "../../db-test-utils";
+import { Iconfile, IconfileDescriptor } from "../../../../src/icon";
+import { query } from "../../../../src/db/db";
+import { boilerplateSubscribe } from "../../../testUtils";
+import { createIcon, deleteIconfile } from "../../../../src/db/repositories/icon-repo";
 
 describe("deleteIconFromDB", () => {
 
@@ -13,16 +14,19 @@ describe("deleteIconFromDB", () => {
 
     it("should delete all entries associated with the icon", done => {
         const user = "zazie";
+        const iconfileDesc: IconfileDescriptor = {
+            format: "french",
+            size: "great"
+        };
         const iconfileInfo: Iconfile = {
             name: "metro-icon",
-            format: "french",
-            size: "great",
+            ...iconfileDesc,
             content: randomBytes(4096)
         };
 
         createIcon(getPool())(iconfileInfo, user)
         .pipe(
-            flatMap(() => deleteIcon(getPool())(iconfileInfo.name, user)),
+            flatMap(() => deleteIconfile(getPool())(iconfileInfo.name, iconfileDesc, user)),
             flatMap(() => query(getPool(), "select count(*) as row_count from icon", [])),
             map(result => expect(result.rows[0].row_count).toEqual("0")),
             flatMap(() => query(getPool(), "select count(*) as row_count from icon_file", [])),
